refactor: move reverse translator into comments
The comments feature owns comment type translation, so the service
doing the reverse lookup belongs with it instead of sitting in the
shared services layer. From there it had to import the default
comment types back out of the slice inside a function body to dodge
an import cycle; after the move a plain import does the job.
