+++ mpvqc/services/__init__.py
@@ -18,7 +18,6 @@ from .player import PlayerService as PlayerService
 from .quit import QuitService as QuitService
 from .resetter import ResetService as ResetService
 from .resource import ResourceService as ResourceService
-from .reverse_translator import ReverseTranslatorService as ReverseTranslatorService
 from .settings import SettingsService as SettingsService
 from .settings_file import SettingsFileService as SettingsFileService
 from .state import StateService as StateService
